add Repository interface
